% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelBuilder.R, R/ModelBuilder_S4.R
\name{build_model}
\alias{build_model}
\alias{build_model,data.frame,character,character-method}
\title{Build a Logistic Regression Model}
\usage{
build_model(dataset, decision_column, model_columns)

build_model(dataset, decision_column, model_columns)

\S4method{build_model}{data.frame,character,character}(dataset, decision_column, model_columns)
}
\arguments{
\item{dataset}{A data frame containing the data to be modeled.}

\item{decision_column}{A string specifying the name of the decision/response column in the dataset.}

\item{model_columns}{A character vector specifying the names of the predictor columns in the dataset.}
}
\value{
A glm object representing the fitted logistic regression model.

An object of class glm representing the fitted logistic regression model.
}
\description{
This function builds a logistic regression model using the specified decision column as the response variable and the model columns as predictors.

Builds a logistic regression model using the specified decision column as the response variable and the model columns as predictors.
}
