% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisHelper.R, R/AnalysisHelper_S4.R
\name{calculate_accuracy}
\alias{calculate_accuracy}
\alias{calculate_accuracy,data.frame-method}
\title{Calculate Accuracy of Predictions}
\usage{
calculate_accuracy(results_dataframe)

calculate_accuracy(results_dataframe)

\S4method{calculate_accuracy}{data.frame}(results_dataframe)
}
\arguments{
\item{results_dataframe}{A data frame with at least two columns: 'ActualResult' and 'PredictedResult'.}
}
\value{
A single numeric value representing the accuracy of the predictions.

A numeric value representing the accuracy of the predictions.
}
\description{
Calculates the accuracy of predicted results against actual results.
}
\keyword{(both}
\keyword{Accuracy}
\keyword{Calculate}
\keyword{Predictions}
\keyword{This}
\keyword{accuracy}
\keyword{actual}
\keyword{against}
\keyword{among}
\keyword{and}
\keyword{as}
\keyword{calculates}
\keyword{cases}
\keyword{defined}
\keyword{examined.}
\keyword{function}
\keyword{internal}
\keyword{is}
\keyword{negatives)}
\keyword{number}
\keyword{of}
\keyword{positives}
\keyword{predicted}
\keyword{proportion}
\keyword{results}
\keyword{results.}
\keyword{the}
\keyword{total}
\keyword{true}
