% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisHelper.R, R/AnalysisHelper_S4.R
\name{generate_output_file}
\alias{generate_output_file}
\alias{generate_output_file,data.frame,data.frame,character,character,character,logical-method}
\title{Generate Output File and Plot}
\usage{
generate_output_file(
  results_dataframe,
  initial_dataset,
  xAxis,
  xAxisLabels,
  decision_column,
  row_names = FALSE
)

generate_output_file(
  results_dataframe,
  initial_dataset,
  xAxis,
  xAxisLabels,
  decision_column,
  row_names = FALSE
)

\S4method{generate_output_file}{data.frame,data.frame,character,character,character,logical}(
  results_dataframe,
  initial_dataset,
  xAxis,
  xAxisLabels,
  decision_column,
  row_names = FALSE
)
}
\arguments{
\item{results_dataframe}{A data frame containing the results to be merged with the initial data set.}

\item{initial_dataset}{A data frame containing the initial data set to be merged with the results.}

\item{xAxis}{The name of the column in `results_dataframe` to be used for the x-axis of the bar plot.}

\item{xAxisLabels}{Labels corresponding to the `xAxis` values, used for annotating the plot.}

\item{decision_column}{The name of the binary decision column in `results_dataframe` used for calculating counts.}

\item{row_names}{Logical value indicating whether to write row names to the output CSV file (default is FALSE).}
}
\value{
The function does not return a value but writes a CSV file and saves a JPEG image of the plot.

NULL The function does not return a value but writes a CSV file and saves a JPEG image of the plot.
}
\description{
This function merges a results data frame with an initial data set, writes the merged data to a CSV file,
and creates a bar plot of the percentage of diabetic cases by a specified category.

Merges a results data frame with an initial data set, writes the merged data to a CSV file,
and creates a bar plot of the percentage of diabetic cases by a specified category.
}
\examples{
# Assuming 'results_dataframe' and 'initial_dataset' are available and have an 'ID' column,
# 'Diabetes_binary' as the decision column, and 'Age_Group' as the category for the x-axis.
# This will create 'output.csv' in the 'data' directory and save a bar plot image 'boxplot.jpeg'.
}
