% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionHelper.R, R/PredictionHelper_S4.R
\name{make_prediction}
\alias{make_prediction}
\alias{make_prediction,data.frame,character,glm,character-method}
\title{Make Predictions Using a Logistic Regression Model}
\usage{
make_prediction(dataset, model_columns, model, decision_column)

make_prediction(dataset, model_columns, model, decision_column)

\S4method{make_prediction}{data.frame,character,glm,character}(dataset, model_columns, model, decision_column)
}
\arguments{
\item{dataset}{A data frame containing the data on which to make predictions.}

\item{model_columns}{A character vector specifying the names of the predictor columns in the dataset.}

\item{model}{An object of class glm representing the fitted logistic regression model.}

\item{decision_column}{A string specifying the name of the actual decision/response column in the dataset.}
}
\value{
A data frame with the original IDs, actual decision column values, and binary predicted results.

A data frame with the original IDs, actual decision column values, and binary predicted results.
}
\description{
This function uses a provided logistic regression model to make predictions on a dataset.
It creates a new data frame containing the IDs, actual decision column values, and predicted results.

Uses a provided logistic regression model to make predictions on a dataset.
}
