% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetUtils.R, R/DatasetUtils_S4.R
\name{read_validate_clean_dataset}
\alias{read_validate_clean_dataset}
\alias{read_validate_clean_dataset,character,character,character,character-method}
\title{Read, Validate, and Optionally Clean a Dataset}
\usage{
read_validate_clean_dataset(
  file_path,
  expected_columns,
  expected_extension,
  na.strings
)

read_validate_clean_dataset(
  file_path,
  expected_columns,
  expected_extension,
  na.strings
)

\S4method{read_validate_clean_dataset}{character,character,character,character}(
  file_path,
  expected_columns,
  expected_extension = "csv",
  na.strings = c("", "NA")
)
}
\arguments{
\item{file_path}{character, the path to the CSV file that should be read.}

\item{expected_columns}{character vector, a list of expected column names.}

\item{expected_extension}{character, the expected file extension (default is "csv").}

\item{na.strings}{character ensure that this code is placed}
}
\value{
A data frame with the validated (and optionally cleaned) dataset.

A data frame with the validated (and optionally cleaned) dataset.
}
\description{
Reads a CSV file from the given file path, checks for the expected columns and file extension,
and validates the presence of an ID column. Optionally, it can also clean the dataset.
}
