% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionManager.R, R/PredictionManager_S4.R
\name{runprediction}
\alias{runprediction}
\alias{runprediction,character,character,character,character,character,character,character-method}
\title{This function orchestrates a prediction pipeline that includes reading and cleaning datasets,
building a model, making predictions, calculating accuracy, and generating output files.}
\usage{
runprediction(
  model_filepath,
  prediction_filepath,
  columns,
  d_column,
  m_columns,
  xAxis_column,
  xAxis_columnlabel
)

runprediction(
  model_filepath,
  prediction_filepath,
  columns,
  d_column,
  m_columns,
  xAxis_column,
  xAxis_columnlabel
)

\S4method{runprediction}{character,character,character,character,character,character,character}(
  model_filepath,
  prediction_filepath,
  columns,
  d_column,
  m_columns,
  xAxis_column,
  xAxis_columnlabel
)
}
\arguments{
\item{model_filepath}{A character string with the file path to the CSV file containing the training data.}

\item{prediction_filepath}{A character string with the file path to the CSV file containing the data for making predictions.}

\item{columns}{A character vector specifying the expected columns in the datasets.}

\item{d_column}{A character string with the name of the decision/response column in the datasets.}

\item{m_columns}{A character vector with the names of the predictor columns in the datasets.}

\item{xAxis_column}{A character string with the name of the column to be used for the x-axis in the output plot.}

\item{xAxis_columnlabel}{A character vector with labels corresponding to the `xAxis_column` values, used for annotating the plot.}
}
\value{
The function does not return a value but prints the accuracy of the model and generates output files.
This will print the accuracy and generate output files including a plot.

NULL The function does not return a value but prints the accuracy of the model and generates output files.
}
\description{
Orchestrates a prediction pipeline including reading and cleaning datasets,
building a model, making predictions, calculating accuracy, and generating output files.
}
